Added caching and isOn method
Added cahcing to PyPHue, which is off by default, but when turned on
could hypothetically cut the load of the phillips hue bridge in half
under the right circumstances. This is important for projects utilizing
PyPHue in a way that requires fast-paced light switches such as my
module PyLights.

Along with the caching feature, I just added an isOn method real quick
that returns whether or not a light is on. This isn't a huge addition,
but seemed like a missing piece of PyPHue.